+++ src/Voximplant.js
@@ -16,7 +16,7 @@ import VideoStream from './call/VideoStream';
 import VideoView from './call/VideoView';
 import EndpointEvents from './call/EndpointEvents';
 import ClientEvents from './client/ClientEvents';
-import {ClientState, CameraType, RenderScaleType, LogLevel, CallError, VideoCodec} from "./Enums";
+import {ClientState, CameraType, RenderScaleType, LogLevel, CallError, VideoCodec, RequestAudioFocusMode} from "./Enums";
 import * as Hardware from './hardware';
 import * as Messaging from './messaging';
 
@@ -56,6 +56,7 @@ export {
     CallError,
     LogLevel,
     VideoCodec,
+    RequestAudioFocusMode,
     Hardware,
     Messaging
 };