+++ src/messaging/Conversation.js
@@ -92,29 +92,63 @@ export default class Conversation {
      */
     isUber;
 
+    /**
+     * @ignore
+     */
     constructor() {
 
     }
 
+    /**
+     * Set the JS object custom data. Note that setting this property does not send changes to the server.
+     * Use the {@link Voximplant.Messaging.Conversation.update()} to send all changes at once.
+     *
+     * @param {object} customData - New custom data of the conversation
+     * @memberOf Voximplant.Messaging.Conversation
+     */
     setCustomData(customData) {
         this.customData = customData;
     }
 
+    /**
+     * Set the distinct flag. Note that setting this property does not send changes to the server.
+     * Use the {@link Voximplant.Messaging.Conversation.update()} to send all changes at once.
+     *
+     * @param {boolean} distinct
+     * @memberOf Voximplant.Messaging.Conversation
+     */
     setDistinct(distinct) {
         this.distinct = distinct;
     }
 
+    /**
+     * Set the public join flag. Note that setting this property does not send changes to the server.
+     * Use the {@link Voximplant.Messaging.Conversation.update()} to send all changes at once.
+     *
+     * @param {boolean} publicJoin
+     * @memberOf Voximplant.Messaging.Conversation
+     */
     setPublicJoin(publicJoin) {
         this.publicJoin = publicJoin;
     }
 
+    /**
+     * Set the conversation title. Note that setting this property does not send changes to the server.
+     * Use the {@link Voximplant.Messaging.Conversation.update()} to send all changes at once.
+     *
+     * @param {string} title
+     * @memberOf Voximplant.Messaging.Conversation
+     */
     setTitle(title) {
         this.title = title;
     }
 
     /**
+     * Add new participants to the conversation. Duplicated users are ignored.
+     * Will fail if any user does not exist. Triggers the MessengerEvents.EditConversation event for all messenger objects on all clients, including this one.
      *
-     * @param participants
+     * @param {Array<ConversationParticipant>} participants - List of ConversationParticipant to be added to the conversation
+     * @memberOf Voximplant.Messaging.Conversation
      */
     addParticipants(participants) {
         if (participants === undefined) {
@@ -124,8 +158,12 @@ export default class Conversation {
     }
 
     /**
+     * Change access rights for the existing participants. This function doesn't apply any changes to the participant list.
+     * Use {@link Voximplant.Messaging.Conversation.addParticipants} or {@link Voximplant.Messaging.Conversation.removeParticipants} instead.
+     * Triggers the MessengerEvents.EditConversation event for all messenger objects on all clients, including this one.
      *
-     * @param participants
+     * @param {Array<ConversationParticipant>} participants - List of ConversationParticipant
+     * @memberOf Voximplant.Messaging.Conversation
      */
     editParticipants(participants) {
         if (participants === undefined) {
@@ -135,8 +173,11 @@ export default class Conversation {
     }
 
     /**
+     * Remove participants from the conversation. Duplicated users are ignored. Will fail if any user does not exist.
+     * Triggers the MessengerEvents.EditConversation event for all messenger objects on all clients, including this one.
      *
-     * @param participants
+     * @param {Array<ConversationParticipant>} participants - List of ConversationParticipant to be removed from the conversation
+     * @memberOf Voximplant.Messaging.Conversation
      */
     removeParticipants(participants) {
         if (participants === undefined) {
@@ -146,23 +187,41 @@ export default class Conversation {
     }
 
     /**
+     * Send conversation changes to the server: title, public join flag, distinct flag and custom data.
+     * Used to send all changes modified via properties.
      *
+     * @memberOf Voximplant.Messaging.Conversation
      */
     update() {
         MessagingModule.updateConversation(this.uuid, this.isUber, this.title, this.publicJoin, this.distinct, this.customData);
     }
 
     /**
+     * Remove current conversation. All participants, including this one, will receive the MessengerEvents.RemoveConversation event.
      *
+     * @memberOf Voximplant.Messaging.Conversation
      */
     remove() {
         MessagingModule.removeConversation(this.uuid);
     }
 
+    /**
+     * Calling this method will inform backend that user is typing some text. Calls within 10s interval from the last call are discarded.
+     *
+     * @memberOf Voximplant.Messaging.Conversation
+     */
     typing() {
         MessagingModule.typing(this.uuid);
     }
 
+    /**
+     * Send message to the conversation. Triggers the MessengerEvents.SendMessage event for all messenger objects on all clients, including this one.
+     *
+     * @param {string} message - Message text
+     * @param {Array<Voximplant.Messaging.Payload>} [payload] - Message payload
+     *
+     * @memberOf Voximplant.Messaging.Conversation
+     */
     sendMessage(message, payload) {
         if (payload === undefined || payload === null) {
             payload = [];
@@ -170,6 +229,14 @@ export default class Conversation {
         MessagingModule.sendMessage(this.uuid, message, payload);
     }
 
+    /**
+     * Mark event as handled by current logged-in device. If single user is logged in on multiple devices,
+     * this can be used to display delivery status by subscribing to the MessengerEvents.Delivered event.
+     *
+     * @param {number} sequence - Sequence number of the event to be marked as delivered
+     *
+     * @memberOf Voximplant.Messaging.Conversation
+     */
     markAsDelivered(sequence) {
         if (sequence === undefined || sequence === null) {
             sequence = 0;
@@ -177,6 +244,14 @@ export default class Conversation {
         MessagingModule.markAsDelivered(this.uuid, sequence);
     }
 
+    /**
+     * Mark the event with the specified sequence as 'read'. This affects 'lastRead' and is used to display unread messages and events.
+     * Triggers the MessengerEvents.Read event for all messenger objects on all connected clients, including this one.
+     *
+     * @param {number} sequence - Sequence number of the event to be marked as read
+     *
+     * @memberOf Voximplant.Messaging.Conversation
+     */
     markAsRead(sequence) {
         if (sequence === undefined || sequence === null) {
             sequence = 0;
@@ -184,6 +259,17 @@ export default class Conversation {
         MessagingModule.markAsRead(this.uuid, sequence);
     }
 
+    /**
+     * Request events in the specified sequence range to be sent from server into this client. Used to get history or
+     * get missed events in case of network disconnect.
+     *
+     * Please note that server will not push any events that was missed due to the client being offline.
+     * Client should use this method to request all events based on the last event sequence received from the server and
+     * last event sequence saved locally (if any).
+     *
+     * @param {number} eventsFrom - First event in range sequence, inclusive
+     * @param {number} eventsTo - First event in range sequence, inclusive
+     */
     retransmitEvents(eventsFrom, eventsTo) {
         // if (eventsFrom === undefined || eventsFrom === null) {
         //     eventsFrom = 0;